:sparkles: atualiza repository

+++ src/main/java/com/senai/biblioteca/repository/EmprestimoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmprestimoRepository extends JpaRepository<EmprestimoEntity, Integer> {
+public interface EmprestimoRepository extends JpaRepository<EmprestimoEntity, Long> {
 }
