+++ src/main/java/com/senai/biblioteca/repository/VisitanteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VisitanteRepository extends JpaRepository<VisitanteEntity, Integer> {
+public interface VisitanteRepository extends JpaRepository<VisitanteEntity, Long> {
 }
